agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.1.0",
    "kafka-python>=2.2.6",
    "requests>=2.31.0",
    "feedparser>=6.0.8",
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.2",
    "neo4j>=5.8.0",
    "pytest>=8.3.5",
    "pyyaml>=6.0",
    "taxii2-client>=2.3.0",
    "prometheus-client>=0.21.1",
    "shodan>=1.31.0",
    "google-generativeai>=0.8.5",
    "dspy>=0.1.4",
    "six>=1.16.0",
    "langdetect>=1.0.9",
    "googletrans>=4.0.2",
    "tenacity>=9.1.2",
    "ratelimit>=2.2.1",
    "boto3>=1.34.70",
    "hvac>=2.1.0",
    "google-cloud-aiplatform>=1.50.0",
    "pydantic>=2.8.0",
    "mitreattack-python>=4.0.2",
    "selectolax>=0.3.29",
    "redis>=4.5.0",
    "jsonschema>=4.17.3",
    "newspaper3k>=0.2.8",
    "readability-lxml>=0.8.4.1",
    "fakeredis>=2.29.0",
    "aiohttp>=3.11.18",
    "aiohttp-retry>=2.9.1",
    "python-dotenv>=1.1.0",
    "pytest-asyncio>=1.0.0",
    "setuptools>=65.0.0",
    "pytest-cov",
    "pytest-forked>=1.6.0",
    "opik>=1.0.0",
    "mistralai>=1.8.2",
    "mmh3>=5.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.1",
    "pytest-mock>=3.12.0",
    "fakeredis>=2.20.1",
    "selectolax>=0.3.29",
    "kafka-python>=2.2.6",
    "prometheus-client>=0.21.1",
]

[tool.uv.sources]

[tool.pytest]
pythonpath = ["."]

[tool.pytest.ini_options]
pythonpath = [
    ".",
    ".."
]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running",
    "e2e: marks tests as end-to-end tests"
]
addopts = "-v --tb=short --strict-markers"
testpaths = ["tests"]
# Filter warnings to reduce noise
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]
